[ObjC]Fix fail() and introduce failWithMessage
The old fail() macro used the `NMB_expectAction` method which
required a block as a parameter, but the macro had no block
argument. This commit fixes the behaviour using a new Objective- C
macro: `failWithMessage` - it allows the user to provide a failure
message. Just like the corresponding Swift method. Calling the
fail() macro resolves in a call to failWithMessage() without a
message to reuse the other macro.

Ref.: #149 #150