Refine ThrowErrorTest testNegativeMatchesWithClosure
The migration to Swift 3 highlighted a small inconsistency in one of the
expected failure message used in the test. That was due to what I think
was a less precise function overload resolution in Swift 2.2, that
resulted in the wrong matcher being called.

This is the rationale behind this change, as described in
https://github.com/norio-nomura/Nimble/pull/5#issuecomment-239262794:

The description of the throwError which is now called by Swift 3 is:

> A Nimble matcher that succeeds when the actual expression throws any error or
> when the passed closures' arbitrary custom matching succeeds.

There is no mention to the type of the error in the description of the matchers
behaviour.

Moreover when I read an expression like

```swift
expect { doStuff() }.to(throwError(closure: closure))
```

I don't read that as an expectation on the type of error, but just on the fact
that it should throw something that satisfy the given closure.

As such I am not expecting any information on the expected error type, because
no such expectation has been set.

On the other hand something like this:

```swift
expect { throw NimbleError.laugh }.to(throwError(errorType: NimbleError.self, closure: closure))
```

I am expecting to have the error type information in the failure message.